db images tg id

+++ src/main/resources/db/changelog/v-1.0/100-user-add-columns-tg.sql
@@ -1,2 +1,2 @@
 alter table users
-    add column tg_id int;
+    add column tg_id bigint;
